Simplify action.yml and enhance entrypoint.sh
Hehehe, alright, so what we did here was, we took out some of the unnecessary stuff from the action.yml file, like the comment severity level, fail on error, and enable gif reactions. I mean, who needs all that, right? 😅

Then, we went ahead and jazzed up the entrypoint.sh file a bit. We're now setting up the aicodebot configuration from the OPENAI_API_KEY and running a code review on the current commit. It's like we're giving it a little extra oomph, you know? 💪

So, in a nutshell, we're making things simpler and more efficient. Just like when I try to fix stuff around the house... except this time it actually works! 😂