Integrate output variables for code review in GitHub workflows
The changes in this commit introduce output variables for the status and comments of the code review in the GitHub workflows. This is achieved by modifying the 'action.yml' and 'entrypoint.sh' files. The former now includes 'review_status' and 'review_comments' as outputs, while the latter has been updated to set these output variables. This enhancement will facilitate the communication of review results in the workflow.