Refactor entrypoint.sh to improve code review feedback
The script `entrypoint.sh` has been refactored to provide more informative feedback during code reviews. The output variables for GitHub workflows have been modified to be written to `$GITHUB_STATE`. Additionally, emojis and comments have been added to the output to indicate the status of the code review. This change enhances the clarity and usefulness of the code review process.