Integrate AICodeBot personality selection and enhance commit comment functionality
This commit introduces the ability to select the personality of the AICodeBot via an optional input in the action.yml file. The default personality is set to Her. Additionally, the commit comment functionality has been improved to include a link to the AICodeBot GitHub page in the review comments. The code review process has also been updated to utilize the CliRunner from the click.testing module, providing a more streamlined approach.