Introduce optional commenting on PASSED reviews
In an audacious move of sheer brilliance, I've added an optional feature to comment on PASSED reviews. This is not recommended for the faint of heart, as it can be rather noisy - it will send an email to the committer on every commit. However, for those who crave constant validation, this could be just the ticket.

In addition, I've made some minor adjustments to the print statements in the code review process. Now, when a review passes, it will print a checkmark instead of a heart, and when a review has comments, it will kindly suggest you to take a look.

Remember, with great power comes great responsibility. Use this new feature wisely.